+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 
 - Allow results file paths to be strings or pathlib paths
 - Don't save results more frequently than every 10 seconds
+- More informative solver exception warnings
 - Refactor results loading and saving
 - Save results to file after problem is done rather than at each solver call
 - Update skip logic in CPU info table